# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Leo Kenzley - Kotlin API
  version: "0.0.1"
servers:
  - url: http://localhost:8080
...
paths:
...
  /users/{id}:
    parameters: # Defina o parâmetro no nível do caminho
      - name: id
        in: path
        description: The id that needs to be fetched. Use user1 for testing
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags:
        - users
      summary: Get user by ID.
      description: Get user detail based on id.
      operationId: getUserById
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid username supplied
        '404':
          description: Resource not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - users
      summary: Update an existing user.
      description: Update an existing user by Id.
      operationId: updateUser
      requestBody:
        description: Update an existing user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '200':
          description: Resource fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid ID format
        '404':
          description: Resource not found
        '422':
          description: Validation error
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - users
      summary: Deletes a user.
      description: Deletes a user based on the id supplied
      operationId: deleteUser
      responses:
        '200':
          description: Resource fetched successfully
        '400':
          description: Invalid ID format
        '404':
          description: Resource not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"            
...
components:
  schemas:
    UserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
          maxLength: 100
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/UserRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
            status:
              type: string
              description: "Status of the user account"
              maxLength: 1
              default: "A"  
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...